Revert "updated go.mod to replace ghedo module with aporeto module (#2)" (#4)
This reverts commit 1af745e7e4c1cf2778e55df3b3628bc6f5a869aa.
We will leave the master to refer to github.com/ghedo so that we
do not pull in this change to any upstream PRs. The change to
the module name will be put on the branch called `go-mod` and this
is the branch that should be referenced from any import statements
in aporeto code.